Add files via upload
This notebook compares three different speech-to-text models (Whisper, Faster Whisper, and Vosk) to determine the most efficient model for audio transcription. The input dataset consists of WAV audio files, which are transcribed using each model. The primary metric used for evaluation is the Word Error Rate (WER). Based on the WER results, the Whisper model was selected as the best performer for this project, as it provided the most accurate transcriptions for the given WAV files.